feat: Easier flow for uploading files to the visualizer
